Some reorganizing
