+++ src/lib/containers/entity/annotations/SchemaDrivenAnnotationEditor.tsx
@@ -23,6 +23,10 @@ import { SynapseClientError } from '../../../utils/SynapseClient'
 import { entityJsonKeys } from '../../../utils/synapseTypes'
 import { SynapseSpinner } from '../../LoadingScreen'
 import { AdditionalPropertiesSchemaField } from './AdditionalPropertiesSchemaField'
+import {
+  dropNullishArrayValues,
+  getFriendlyPropertyName,
+} from './AnnotationEditorUtils'
 import { CustomAdditionalPropertiesFieldTemplate } from './CustomAdditionalPropertiesFieldTemplate'
 import { CustomArrayFieldTemplate } from './CustomArrayFieldTemplate'
 import { CustomBooleanWidget } from './CustomBooleanWidget'
@@ -47,41 +51,6 @@ export type SchemaDrivenAnnotationEditorModalProps = {
   entityId: string
   show: boolean
   onHide: () => void
-}
-
-/**
- * Strips null values from arrays in the provided form data. If the array is empty after
- * removing null values, the key is removed from the form data.
- *
- * This allows users to submit forms with empty array fields (SWC-5762)
- */
-function dropNullishArrayValues(
-  formData: Record<string, unknown>,
-): Record<string, unknown> {
-  const newFormData: Record<string, unknown> = {}
-  Object.keys(formData).forEach(key => {
-    let value = formData[key]
-    if (Array.isArray(value)) {
-      value = (value as Array<any>).filter((item: any) => item != null)
-      if (!isEmpty(value)) {
-        newFormData[key] = value
-      }
-    } else {
-      newFormData[key] = value
-    }
-  })
-  return newFormData
-}
-
-function getFriendlyPropertyName(error: AjvError) {
-  if (error.property.startsWith('[')) {
-    // Additional properties are surrounded by brackets and quotations, so let's remove them
-    return error.property.substring(2, error.property.length - 2)
-  } else if (error.property.startsWith('.')) {
-    return error.property.substring(1)
-  } else {
-    return error.property
-  }
 }
 
 // patternProperties lets us define how to treat additionalProperties in a JSON schema by property name
@@ -232,7 +201,7 @@ export const SchemaDrivenAnnotationEditor: React.FunctionComponent<SchemaDrivenA
               variant="info"
               transition={false}
             >
-              <b>{entityJson.name as string}</b> has no annotations. Click the{' '}
+              <b>{entityJson.name}</b> has no annotations. Click the{' '}
               <AddToList /> button to annotate.
             </Alert>
           )}
